pyproject.toml
[tool.poetry]
name = "pieces_copilot_sdk"
version = "0.6.0"
description = "A Pieces OS SDK wrapper that allows anyone to easily create a Pieces Copilot experience with minimal code."
authors = ["Mason Williams <mason@pieces.app>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
pieces_os_client = "^2.7.0"
websocket-client = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
